Simplify dependency management
