+++ buildSrc/build.gradle.kts
@@ -24,8 +24,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+buildscript {
+    repositories {
+        maven("https://plugins.gradle.org/m2/")
+    }
+
+    dependencies {
+        classpath("com.github.jk1:gradle-license-report:1.16")
+    }
+}
+
 plugins {
+    java
     `kotlin-dsl`
+    id("com.github.jk1.dependency-license-report") version "1.16"
 }
 
 kotlinDslPluginOptions {
@@ -34,11 +46,14 @@ kotlinDslPluginOptions {
 
 repositories {
     mavenLocal()
+    maven("https://plugins.gradle.org/m2/")
     mavenCentral()
 }
 
 val jacksonVersion = "2.11.0"
+val licenseReportVersion = "1.16"
 
 dependencies {
     implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-xml:$jacksonVersion")
+    api("com.github.jk1:gradle-license-report:${licenseReportVersion}")
 }
